Move slots types to a separate file to simplify files with logic.
